Replace LOGGER by LOG
Replace LOGGER by LOG to follow the
OpenDaylight recommendations [1].

[1]
https://wiki.opendaylight.org/view/BestPractices/Logging_Best_Practices

+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/VtyshCli.java
@@ -175,8 +175,8 @@ public class VtyshCli extends OsgiCommandSupport {
                 socket.close();
                 return;
             }
-            if ((ip == (int) gt) || (ip == (int) hashChar)) {
-                if (ip == (int) gt) {
+            if (ip == gt || ip == hashChar) {
+                if (ip == gt) {
                     sb.append(gt);
                 } else {
                     sb.append(hashChar);
@@ -207,7 +207,7 @@ public class VtyshCli extends OsgiCommandSupport {
                 socket.close();
                 return;
             }
-            if ((ip == (int) gt) || (ip == (int) hashChar)) {
+            if (ip == gt || ip == hashChar) {
                 break;
             } else if (ip == -1) {
                 session.getConsole().println(sb.toString());
